added linter for code block line length and gen_numbering (#70)
## Improvements:

- Code Blocks shall only have a max line length of 100 characters,
because otherwise they are not completely included in the syllabus pdf.
This is now automatically checked during build and build will fail if a
line is longer.
- The Python script to generate the chapter numbers to the headings, has
been improved, to return with an exit code != 0 in case of a fix. This
will now be used as check in the build and also available as pre-commit
hook to the maintainers.

---------

Signed-off-by: René <snooz@posteo.de>